requirements.txt
Django>=2.0
django-ipware>=1.1.6
djangorestframework>=3.7.7
psycopg2-binary>=2.7.4
python-dateutil>=2.6.1
pytz>=2017.3
requests>=2.18.4
setuptools>=36.6.0
sendgrid==5.3.0
